requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=eigen
absl-py=0.11.0=pyhd3eb1b0_1
aiohttp=3.7.3=py38h2bbff1b_1
astunparse=1.6.3=py_0
async-timeout=3.0.1=py38haa95532_0
attrs=20.3.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py38haa95532_0
blosc=1.20.1=h7bd577a_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py38h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2021.1.19=haa95532_0
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2020.12.5=py38haa95532_0
cffi=1.14.4=py38hcd4344a_0
chardet=3.0.4=py38haa95532_1003
charls=2.1.0=h33f27b4_2
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
cryptography=2.9.2=py38h7a1dbc1_0
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38he774522_0
dask-core=2021.1.1=pyhd3eb1b0_0
decorator=4.4.2=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
gast=0.4.0=py_0
geos=3.8.0=h33f27b4_0
giflib=5.2.1=h62dcd97_0
google-auth=1.24.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
google-pasta=0.2.0=py_0
grpcio=1.31.0=py38he7da953_0
h5py=2.10.0=py38h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
imagecodecs=2021.1.11=py38h5da4933_1
imageio=2.9.0=py_0
imgaug=0.4.0=pyhd3eb1b0_0
importlib-metadata=2.0.0=py_1
intel-openmp=2020.2=254
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py38hd77b12b_0
lcms2=2.11=hc51a39a_0
lerc=2.2.1=hd77b12b_0
libaec=1.0.4=h33f27b4_1
libdeflate=1.7=h2bbff1b_5
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h23ce68f_0
libtiff=4.1.0=h56a325e_1
libzopfli=1.0.3=ha925a31_0
lz4-c=1.9.3=h2bbff1b_0
markdown=3.3.3=py38haa95532_0
matplotlib=3.3.2=haa95532_0
matplotlib-base=3.3.2=py38hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.2.0=py38h45dec08_0
mkl_random=1.1.1=py38h47e9c7a_0
multidict=4.7.6=py38he774522_1
networkx=2.5=py_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
oauthlib=3.1.0=py_0
olefile=0.46=py_0
opencv=4.0.1=py38h2a7c758_0
openjpeg=2.3.0=h5ec785f_1
openssl=1.1.1i=h2bbff1b_0
opt_einsum=3.1.0=py_0
pandas=1.2.1=py38hf11a4ad_0
pillow=8.1.0=py38h4fa10fc_0
pip=20.3.3=py38haa95532_0
protobuf=3.14.0=py38hd77b12b_1
py-opencv=4.0.1=py38he44ac1e_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyjwt=2.0.1=py38haa95532_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38ha925a31_4
pyreadline=2.1=py38_1
pysocks=1.7.1=py38haa95532_0
python=3.8.5=h5fd99cc_1
python-dateutil=2.8.1=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py38he774522_2
pyyaml=5.4.1=py38h2bbff1b_1
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7=pyhd3eb1b0_1
scikit-image=0.17.2=py38h1e1f486_0
scikit-learn=0.23.2=py38h47e9c7a_0
scipy=1.6.0=py38h14eb087_0
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=52.0.0=py38haa95532_0
shapely=1.7.1=py38h210f175_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py38haa95532_0
snappy=1.1.8=h33f27b4_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.3.0=mkl_py38h8c0d9a2_0
tensorflow-base=2.3.0=eigen_py38h75a453f_0
tensorflow-estimator=2.3.0=pyheb71bc4_0
termcolor=1.1.0=py38_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.1.14=pyhd3eb1b0_1
tk=8.6.10=he774522_0
toolz=0.11.1=pyhd3eb1b0_0
tornado=6.1=py38h2bbff1b_0
tqdm=4.56.0=pyhd3eb1b0_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.3=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38_0
wrapt=1.12.1=py38he774522_1
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
yarl=1.5.1=py38he774522_0
zfp=0.5.5=hd77b12b_4
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
